UserId long, user-page, refactored ad navigation, disabled url logging, login- transactionlogging, User-Bean added

+++ src/main/java/de/hh/changeRing/domain/Transaction.java
@@ -36,10 +36,10 @@ public class Transaction {
     @XmlElement
     private Date date;
     @XmlElement
-    private int fromId;
+    private Long fromId;
     private User from;
     @XmlElement
-    private int toId;
+    private Long toId;
     private User to;
     @XmlElement
     private long amount;
